refactor: rename devcontainer package to devsandbox
N
