+++ builder-stack/cnb-builder-shim/internal/devsandbox/devsandbox_suite_test.go
@@ -16,7 +16,7 @@
  * to the current version of the project delivered to anyone in the future.
  */
 
-package devcontainer_test
+package devsandbox_test
 
 import (
 	"testing"
