+++ builder-stack/cnb-builder-shim/internal/devsandbox/webserver/service/testdata/django-helloworld.zip
